Missing snooze screen but thats okay.

+++ app/src/main/java/com/linklab/emmanuelogunjirin/besi_c/EndOfDayEMA.java
@@ -37,7 +37,7 @@ public class EndOfDayEMA extends WearableActivity       // This is the main acti
 
     private Timer EMARemindertimer;
     private int EMAReminderDelay = new Preferences().PainEMAReminderDelay;
-    private int EMAReminderInterval = new Preferences().PainEMAReminderInterval; //Time before pinging user after not finishing EMA
+    private long EMAReminderInterval = new Preferences().PainEMAReminderInterval; //Time before pinging user after not finishing EMA
     private int ReminderNumber = new Preferences().PainEMAReminderNumber;
     private int ReminderCount = 0;
     private int CurrentQuestion = 0;
